(PC-18688) feat(useGetDepositAmountsByAge): create a user feature and move useGetDepositAmountsByAge helper in it
