+++ src/features/home/components/headers/HomeHeader.tsx
@@ -9,8 +9,8 @@ import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBene
 import { GeolocationBanner } from 'features/home/components/GeolocationBanner'
 import { useAvailableCredit } from 'features/home/services/useAvailableCredit'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { useGetDepositAmountsByAge } from 'features/offer/helpers/useGetDepositAmountsByAge/useGetDepositAmountsByAge'
 import { isUserBeneficiary } from 'features/profile/utils'
+import { useGetDepositAmountsByAge } from 'features/user/helpers/useGetDepositAmountsByAge/useGetDepositAmountsByAge'
 import { env } from 'libs/environment'
 import { useGeolocation, GeolocPermissionState } from 'libs/geolocation'
 import { formatToFrenchDecimal } from 'libs/parsers'
